fix: split did::mock to make benchmarks work again
