+++ pallets/did/src/benchmarking.rs
@@ -27,7 +27,7 @@ use sp_runtime::{traits::IdentifyAccount, MultiSigner, SaturatedConversion};
 
 use crate::{
 	did_details::*,
-	mock::{
+	mock_utils::{
 		generate_base_did_creation_details, generate_base_did_details, generate_base_did_update_details,
 		get_key_agreement_keys, get_public_keys, get_service_endpoints, DEFAULT_URL_SCHEME,
 	},
@@ -488,6 +488,6 @@ benchmarks! {
 
 impl_benchmark_test_suite! {
 	Pallet,
-	crate::mock::std::ExtBuilder::default().build_with_keystore(None),
-	crate::mock::std::Test
+	crate::mock::ExtBuilder::default().build_with_keystore(None),
+	crate::mock::Test
 }
